(maint)) Refactoring
